:sparkles: update security in pages
