+++ app/_components/ErrorPage.tsx
@@ -1,0 +1,13 @@
+
+import ReturnToLoginPage from "./ReturnToLoginPage";
+
+export default function ErrorPage() {
+
+    return (
+        <main>
+            잘못된 접근입니다. 돌아가세요
+            <ReturnToLoginPage/>
+        </main>
+    )
+
+}
