+++ app/dashboard/page.tsx
@@ -2,24 +2,10 @@
 import TodaySportSection from "./_sections/TodaySportSection";
 import Calendar from "./_components/Calendar";
 import MenuBar from "@/app/_components/MenuBar";
-import {cookies} from "next/headers";
 
-import ReturnToLoginPage from "./_components/ReturnToLoginPage";
+import ErrorPage from "@/app/_components/ErrorPage";
 
 export default async function MainPage() {
-
-  const cookieStore = await cookies();
-  const uToken = cookieStore.get('uToken');
-  console.log(uToken);
-
-  if (uToken == null) {
-    return (
-        <main>
-            잘못된 접근입니다. 돌아가세요
-            <ReturnToLoginPage/>
-        </main>
-    )
-  }
 
   return (
     <main className="flex min-h-screen flex-col bg-white gap-5">
