+++ app/kakao-signup/_components/KakaoSignUpForm.tsx
@@ -23,6 +23,7 @@ const KakaoSignUpForm = () => {
                 headers: {
                     "Content-Type": "application/json",
                 },
+                credentials: 'include',
                 body: JSON.stringify({
                     name,
                     nickName,
