+++ app/kakao-signup/page.tsx
@@ -2,13 +2,20 @@
 import KakaoSignUpForm from "./_components/KakaoSignUpForm";
 import Link from "next/link";
 import Image from "next/image";
-// import { cookies } from 'next/headers';
+import { cookies } from 'next/headers';
 
-export default function SignUpPage(){
+import ErrorPage from "@/app/_components/ErrorPage";
 
-//     console.log(cookies());
-//     const email = cookies().get('email')?.value;
-//     console.log(email);
+export default async function SignUpPage(){
+
+    const cookieStore = await cookies();
+    const JSESSIONID = cookieStore.get('JSESSIONID');
+    if (JSESSIONID == null) {
+        return (
+            <ErrorPage/>
+        )
+    }
+
     const apiUrl = process.env.NEXT_PUBLIC_API_URL;
     const url = `${apiUrl}/kakao`
