+++ app/write-diary/page.tsx
@@ -1,9 +1,20 @@
 
 import WriteForm from "./_component/WriteForm";
 import BackButton from "@/app/_components/BackButton";
+import {cookies} from "next/headers";
 
-export default function WriteDiary() {
+import ErrorPage from "@/app/_components/ErrorPage";
 
+export default async function WriteDiary() {
+
+  const cookieStore = await cookies();
+  const JSESSIONID = cookieStore.get('JSESSIONID');
+
+  if (JSESSIONID == null) {
+    return (
+      <ErrorPage/>
+    )
+  }
 
     return (
         <main>
